Add files via upload
This is a basic Next Word Predictor, made on a small dataset collected from a You tube channel on ML. The code shows how one can basically train a model to predict the subsequent words in a sentence. This project is under development, just the most basic version is uploaded here.